Split CDS viewer state into separate file. Add visible layers bound setting to scatter viewer.
